Rename channel pages/components
